+++ CHANGELOG.md
@@ -7,14 +7,26 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ---
 
-## [1.2.1] - 2025-07-27
+## [v1.2.2] - 2025-07-27
+
+### Added
+- Output message for in-place encrypt / decrypt
+- Tests for in-place mode of operation
+- Additional tests for -c and -i
+
+### Changed
+- Restructure & rename test methods for clarity / maintainability
+
+---
+
+## [v1.2.1] - 2025-07-27
 
 ### Fixed
 - Modified test_cases to clearly indicate that data is for testing purposes only
 
 ---
 
-## [1.2.0] - 2025-07-27
+## [v1.2.0] - 2025-07-27
 
 ### Added
 - Added `-v` flag to display version info
@@ -29,7 +41,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ---
 
-## [1.1.0] - 2025-07-25
+## [v1.1.0] - 2025-07-25
 
 ### Added
 - Add ability to specify a custom key file path via `SCCRYPT_KEY_FILE` environment variable
@@ -40,7 +52,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Improve usage info and error messages
 - Update documentation
 
-## [1.0.0] - 2025-07-24
+## [v1.0.0] - 2025-07-24
 
 ### Added
 - Initial release of sccrypt